pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dm-agent"
version = "0.1.0"
description = "HPC Data Management Agent - automated scanning, analysis, and cleanup"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
api = [
    "anthropic>=0.39.0",
]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
dm-agent = "dm_agent.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
